validacao do login e pass e guardar informaçoes na business database

+++ CPTWorkouts/Models/Faturas.cs
@@ -16,10 +16,16 @@ namespace CPTWorkouts.Models
         /// Chave estrangeira correspondente aos serviços vendidos ao Cliente
         ///</summary>
         public string serviçosFK { get; set; }
+        /// <summary>
+        /// auxiliary attribute to help us to write fee values
+        /// como é string aceita qualquer valor, temos de escrever uma expressão regular para regular isso
+        /// </summary>
+        /// 
+        /// 
         ///<summary>
         /// Valor de cada serviço adquirido pelo cliente
         ///</summary>
-        public int Valor {  get; set; }
+        public decimal Valor {  get; set; }
 
         // relação 1-1 com Clientes
         public Clientes Cliente { get; set; }
